# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ISPO Smart Schedule API
  version: 1.0.0
  description: |
    Backend API для системы "Умное расписание" (ИСПО СПбПУ).
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and get new JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/schedule/current:
    get:
      tags: [Schedule]
      summary: Current week schedule
      parameters:
        - in: query
          name: group_id
          required: true
          schema: { type: integer, minimum: 1 }
        - in: query
          name: date
          required: false
          description: "Reference date (YYYY-MM-DD). Default: today (UTC)"
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/schedule/range:
    get:
      tags: [Schedule]
      summary: Schedule for date range
      parameters:
        - in: query
          name: group_id
          required: true
          schema: { type: integer, minimum: 1 }
        - in: query
          name: date_start
          required: true
          schema: { type: string, format: date }
        - in: query
          name: date_end
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/schedule/pdf:
    get:
      tags: [Schedule]
      summary: PDF for two weeks (week1+week2)
      parameters:
        - in: query
          name: group_id
          required: true
          schema: { type: integer, minimum: 1 }
        - in: query
          name: date_start
          required: true
          description: Any date in week1 (YYYY-MM-DD)
          schema: { type: string, format: date }
      responses:
        "200":
          description: PDF document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/push/register:
    post:
      tags: [Push]
      summary: Register device token for group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushRegisterRequest"
      responses:
        "201":
          description: Registered
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/push/unregister:
    post:
      tags: [Push]
      summary: Unregister device token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnregisterRequest"
      responses:
        "204":
          description: Unregistered
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/groups:
    get:
      tags: [Dictionaries]
      summary: List groups (public)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Group" }
  /api/v1/subjects:
    get:
      tags: [Dictionaries]
      summary: List subjects (public)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Subject" }
  /api/v1/locations:
    get:
      tags: [Dictionaries]
      summary: List locations (public)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Location" }
  /api/v1/campuses:
    get:
      tags: [Dictionaries]
      summary: List campuses (public)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Campus" }
  /api/v1/location-types:
    get:
      tags: [Dictionaries]
      summary: List location types (public)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LocationType" }
  /api/v1/admin/groups:
    get:
      tags: [Admin]
      summary: List groups
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Group" }
    post:
      tags: [Admin]
      summary: Create group (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Group" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
  /api/v1/admin/groups/{id}:
    put:
      tags: [Admin]
      summary: Update group (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Group" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
    delete:
      tags: [Admin]
      summary: Delete group (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Deleted }
  /api/v1/admin/subjects:
    get:
      tags: [Admin]
      summary: List subjects
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Subject" }
    post:
      tags: [Admin]
      summary: Create subject (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Subject" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subject" }
  /api/v1/admin/subjects/{id}:
    put:
      tags: [Admin]
      summary: Update subject (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Subject" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subject" }
    delete:
      tags: [Admin]
      summary: Delete subject (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Deleted }
  /api/v1/admin/locations:
    get:
      tags: [Admin]
      summary: List locations
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Location" }
    post:
      tags: [Admin]
      summary: Create location (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Location" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
  /api/v1/admin/locations/{id}:
    put:
      tags: [Admin]
      summary: Update location (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Location" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
    delete:
      tags: [Admin]
      summary: Delete location (admin only)
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Deleted }
  /api/v1/admin/campuses:
    get:
      tags: [Admin]
      summary: List campuses
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Campus" }
    post:
      tags: [Admin]
      summary: Create campus
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Campus" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campus" }
  /api/v1/admin/campuses/{id}:
    put:
      tags: [Admin]
      summary: Update campus
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Campus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campus" }
    delete:
      tags: [Admin]
      summary: Delete campus
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Deleted }
  /api/v1/admin/location-types:
    get:
      tags: [Admin]
      summary: List location types
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LocationType" }
    post:
      tags: [Admin]
      summary: Create location type
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocationType" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocationType" }
...
  /api/v1/admin/location-type-links/{location_id}/{type_id}:
    delete:
      tags: [Admin]
      summary: Delete location/type link
      security: [{ bearerAuth: [] }, { adminKey: [] }]
      parameters:
        - in: path
          name: location_id
          required: true
          schema: { type: integer, minimum: 1 }
        - in: path
          name: type_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      required:
        [access_token, expires_at, refresh_token, refresh_expires_at, role]
      properties:
        access_token: { type: string }
        expires_at: { type: string, format: date-time }
        refresh_token: { type: string }
        refresh_expires_at: { type: string, format: date-time }
        role:
          type: string
          enum: [student, dispatcher, admin, viewer]
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    RefreshResponse:
      type: object
      required: [access_token, expires_at, refresh_token, refresh_expires_at]
      properties:
        access_token: { type: string }
        expires_at: { type: string, format: date-time }
        refresh_token: { type: string }
        refresh_expires_at: { type: string, format: date-time }
...
    PushRegisterRequest:
      type: object
      required: [group_id, token]
      properties:
        group_id: { type: integer, minimum: 1 }
        token: { type: string }
...